Moved string array to different files.
