+++ CMakeLists.txt
@@ -1,6 +1,7 @@
 # The minimum required version hasn't been tested. Feel free to adjust
 # downwards if necessary.
 cmake_minimum_required(VERSION 2.8)
+set (CMAKE_BUILD_TYPE Debug)
 
 project(flyby C)
 
@@ -24,7 +25,7 @@ set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wall")
 
 configure_file(config.h.in config.h @ONLY)
 include_directories(${CMAKE_BINARY_DIR})
-add_executable(flyby flyby_ui.c flyby_hamlib.c flyby_main.c)
+add_executable(flyby flyby_ui.c flyby_hamlib.c flyby_main.c string_array.c)
 install(TARGETS flyby RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR})
 
 target_link_libraries(flyby m)
